requirements.txt
python-dotenv
openai
flask
psycopg2-binary
requests
scikit-learn
